refactor: 경험 보드 순서 변경 로직 분리
